factory: build pinephone factory img
Add a script to create the three-layered factory image for the
PinePhone, based on Martijn's factorytest. factorytest was also used for
the Brave Heart and to flash the ubports CE. Add factorytest and related
pmaports, as well as dependencies that are only in Alpine testing (so we
can build the layer 3 image with v20.05 instead of edge). There is a lot
that could be optimized here, but that doesn't seem like a good use of
time.

Move some code from build.sh to common.sh, so it can be used by the new
script.